alias in tsconfig.app. rename IRoute. routes+index in NavLinks
