+++ src/common/components/Navigation/NavigationLinks.tsx
@@ -1,23 +1,20 @@
-import { IRoute } from "@common/types/IRoute";
 import { Link } from "react-router";
 import MenuItem from "@mui/material/MenuItem";
 import Typography from "@mui/material/Typography";
-
+import { routes } from "@common/constants.tsx";
 function NavigationLinks({
-  routes,
   isSettings,
   keySuffix,
   handleCloseNavMenu,
 }: {
-  routes: IRoute[];
   isSettings: boolean;
   keySuffix: string;
   handleCloseNavMenu: () => void;
 }) {
   return routes
     .filter((page) => page.isSettings === isSettings)
-    .map((page) => (
-      <Link key={page.page + "_" + keySuffix} to={page.href}>
+    .map((page,index) => (
+      <Link key={index + "_" + keySuffix} to={page.href}>
         <MenuItem onClick={handleCloseNavMenu}>
           <Typography sx={{ textAlign: "center" }}>{page.page}</Typography>
         </MenuItem>
