+++ src/common/constants.tsx
@@ -1,8 +1,8 @@
-import Dashboard from "../pages/Dashboard/Dashboard.tsx";
-import Generator from "../pages/Generator/Generator.tsx";
-import Settings from "../pages/Settings/Settings.tsx";
-import Login from "../pages/Login/Login.tsx";
-import { IRoute } from "@common/types/IRoute.tsx";
+import Dashboard from "@pages/Dashboard/Dashboard.tsx";
+import Generator from "@pages/Generator/Generator.tsx";
+import Settings from "@pages/Settings/Settings.tsx";
+import Login from "@pages/Login/Login.tsx";
+import { IRoute } from "@common/types/Route.tsx";
 
 export const routes: IRoute[] = [
   { page: "Главная", href: "/", isSettings: false, component: <Dashboard /> },
