+++ tsconfig.app.json
@@ -22,6 +22,13 @@
     "noFallthroughCasesInSwitch": true,
     "noUncheckedSideEffectImports": true,
 
+    //absolute import
+    "baseUrl": "./",
+    "paths": {
+      "@assets/*": ["./src/assets/*"] ,
+      "@common/*": ["./src/common/*"],
+      "@pages/*": ["./src/pages/*"],
+    }
   },
   "include": ["src"]
 }
